% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{add_class}
\alias{add_class}
\alias{prepend_class}
\alias{append_class}
\title{Add (S3) classes to an object}
\usage{
prepend_class(x, cls = character(0), null = list())

append_class(x, cls = character(0), null = list())
}
\arguments{
\item{x}{The object whose class is to be updated.}

\item{cls}{The class(es) to add.}

\item{null}{If \code{x} is \code{NULL}, first replace it with this value, since \code{NULL} can't have a class.}
}
\value{
The object with new class structure.
}
\description{
Add (S3) classes to an object
}
\details{
The new class vector will be uniquified, in left-to-right preferential order (so \verb{x,y,x,z} becomes \verb{x,y,z}).
}
