% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_ifnot}
\alias{default_ifnot}
\alias{default_if_not}
\title{Predicate-based object replacement}
\usage{
default_ifnot(x, test = function(x) TRUE, default = x)

default_if_not(x, test = function(x) TRUE, default = x)
}
\description{
Predicate-based object replacement
}
\details{
Meant to be pipe-friendly and similar in spirit to \code{tidyr::replace_na}, only:
(i) this function is scalar, operating on the entire object as a single unit and
(ii) one passes an arbitrary predicate function \code{test} that takes a single argument (\code{x}).
}
