% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.core-package.R
\docType{package}
\name{murlib.core-package}
\alias{murlib.core}
\alias{murlib.core-package}
\title{murlib.core: Core Utilities}
\description{
A collection of core utilities that I use frequently that are missing from 'core' R.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
