% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{replace_na}
\alias{replace_na}
\alias{na2}
\alias{nan2}
\alias{inf2}
\alias{na_to}
\alias{nan_to}
\alias{inf_to}
\title{Replace specials (NA, NaN, (+/-)Inf)}
\usage{
na2(x, replacement)

nan2(x, replacement)

inf2(x, replacement)

na_to(x, replacement)

nan_to(x, replacement)

inf_to(x, replacement)
}
\description{
Replace specials (NA, NaN, (+/-)Inf)
}
