% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{safe_sample}
\alias{safe_sample}
\alias{sample_int}
\alias{sample_from}
\title{Sample without the \emph{n}=1 gotcha}
\usage{
sample_int(
  n,
  size = n,
  replace = FALSE,
  prob = NULL,
  useHash = (n > 1e+07 && !replace && is.null(prob) && size <= n/2)
)

sample_from(x, size = length(x), replace = FALSE, prob = NULL)
}
\description{
Sample without the \emph{n}=1 gotcha
}
