% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\docType{data}
\name{typed_na}
\alias{typed_na}
\alias{na_lgl}
\alias{na_int}
\alias{na_dbl}
\alias{na_chr}
\alias{na_date}
\alias{na_posixct}
\title{Typed NAs}
\format{
An object of class \code{logical} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 1.

An object of class \code{Date} of length 1.

An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.
}
\usage{
na_lgl

na_int

na_dbl

na_chr

na_date

na_posixct
}
\description{
Typed NAs
}
\keyword{datasets}
